WIP(1) Made separate modules for each standalone class. Also made few code style improvements.
