+++ card.py
@@ -1,0 +1,8 @@
+# coding: utf-8
+
+
+class Card:
+    def __init__(self, name, colour):
+        self.name = name
+        self.colour = colour
+
