+++ disease.py
@@ -1,0 +1,8 @@
+# coding: utf-8
+
+
+class Disease:
+    def __init__(self, colour):
+        self.colour = colour
+        self.cured = False
+
